Split station into two panels and add more sound effects
Implemented the dock sound effect and added a transaction sound effect. Known issues with the current multi-panel system include adjustment of the selection if the player waits and new ships dock, inability to loop around each list, and switching columns moving the selection to the top. Discovered a bug with ship's updatePrices(), which makes it update expander prices for ALL ships.