Remove sector discovery system and add initial galaxy features
Instead of keeping a record of discovered sectors, only the current FOV is shown, which has been expanded to 3 tiles. At the galaxy sizes tested, localized rendering has not be necessary, though it can be easily added. Scanners can no longer scan the map, though it is planned for them to increase the FOV radius.