Allow for travel between galaxies
Leaving a galaxy currently costs the same as a normal burn between sectors. More testing is needed to make sure that no crashes occur while the player is not present in their destination galaxy. Fixed a bug where sound volume wouldn't be set correctly when loading the game. Fixed a crash when unaligned ships choose aligned targets to attack. Fixed a crash when ships flee a battle.