Rework the name system to be purely on a per-game basis
Names are stored for the player's ship and its captain. The ship's name is used in-game when the player's ship is referred to. The captain's name is saved in the high score log. Both prompts can be skipped by pressing enter and a default name will be used. Fixes options not being saved if the ship is destroyed later in the game.