Fix jar manifest
Fixes manifest generation and generates a "fat" jar, removing the need to bundle libraries. This does mess with the path, so Main will automatically detect IDE compilation and updates the path accordingly. Also changes the exception handler to print stack traces to System.out, for debugging from an IDE or through the command line.