Overhaul project architecture for maintainability
This commit started as an effort to eliminate this project's dependency
on a compiled JAR of APWT. To do this, build.gradle was updated to pull
in APWT and the other necessary libraries from JitPack. The Gradle
version used by the wrapper was also bumped to 7.3.3.

It turns out that I had pushed some breaking changes to the APWT
repository that had not made their way into the JAR that EverSector was
using. Thus, many files had to be updated for compatibility with this
new version (mainly changing the utility package name and using two ints
instead of a SquidLib Coord for display coordinates).

Furthermore, the Splash class that EverSector used for nebula generation
was deprecated in favor of the Spill class as of SquidLib v3.0.0. Galaxy
generation was updated to use the potentially more efficient MultiSpill
class. Nebula size and count were also tweaked slightly, which should
improve the quality and appearance of generated galaxies.

It seems that the updates to how the project were built also
necessitated some changes to the building and packaging infrastructure
of the project. Instead of using a custom bundle directory into which
JARs are copied, the distribution generated by Gradle is used as the
base into which other files (e.g. assets) are copied. To simplify this
process, I wrote a new build.sh script.

For ease of access, I merged the two READMEs (bundle and repository). I
also added credits for the base tilesets used in the game. (Apologies to
the original tileset creators for the delay!) I also decided to include
the audio in the repository, since it is not otherwise stored in a
robust and accessible place. However, due to the size of the soundtrack
(over 100 MB as a 10 minute uncompressed wav file), it was not able to
be pushed to GitHub without LFS. As a workaround, I've re-encoded it at
a 32kHz sample rate, which drops the file size and sounds about the same
to me.

Apparently, AsciiPanel uses the Java class loader to load font image
resources, meaning the asset paths must be on the classpath. This
deserves an upstream fix, but for the time being, I have added the base
directory of the bundle to the classpath and used relative paths for
font loading.

All these changes should make it possible to build EverSector more
reliably with modern tools. Currently I have only tested these changes
on Linux, and Windows support is a lower priority. I do not intend to a
bundled JRE version for Windows as I have in the past, mainly due to all
the headaches it caused. In any case, I do intend to release a new
version with these changes in the near future!